Disable uniform bound

+++ src/project/ProjectSettings.cpp
@@ -58,12 +58,16 @@ ProjectSettings::~ProjectSettings()
 {
 
 }
-
+/**
+ * @brief ProjectSettings::fromJson, here we fix tmin and tmax as integer value
+ * @param json
+ * @return
+ */
 ProjectSettings ProjectSettings::fromJson(const QJsonObject& json)
 {
     ProjectSettings settings;
-    settings.mTmin = (double) json.contains(STATE_SETTINGS_TMIN) ? json.value(STATE_SETTINGS_TMIN).toInt() : STATE_SETTINGS_TMIN_DEF;
-    settings.mTmax = (double) json.contains(STATE_SETTINGS_TMAX) ? json.value(STATE_SETTINGS_TMAX).toInt() : STATE_SETTINGS_TMAX_DEF;
+    settings.mTmin = json.contains(STATE_SETTINGS_TMIN) ? json.value(STATE_SETTINGS_TMIN).toInt() : STATE_SETTINGS_TMIN_DEF;
+    settings.mTmax = json.contains(STATE_SETTINGS_TMAX) ? json.value(STATE_SETTINGS_TMAX).toInt() : STATE_SETTINGS_TMAX_DEF;
     settings.mStep = json.contains(STATE_SETTINGS_STEP) ? json.value(STATE_SETTINGS_STEP).toDouble() : STATE_SETTINGS_STEP_DEF;
     settings.mStepForced = json.contains(STATE_SETTINGS_STEP_FORCED) ? json.value(STATE_SETTINGS_STEP_FORCED).toBool() : STATE_SETTINGS_STEP_FORCED_DEF;
